# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Auspac Documentation",
        "description": "Project Inventory Auspac Laravel 12 Documentation",
        "contact": {
            "email": "example@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/xero/auth": {
            "get": {
                "tags": [
                    "Xero"
                ],
                "summary": "Redirect to Xero authorization screen",
                "description": "Redirects the user to Xero's OAuth2 login and consent screen.",
                "operationId": "c18ccef982c829e6b18a7a5084d92b67",
                "responses": {
                    "302": {
                        "description": "Redirect to Xero authorization URL"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
        "/api/xero/callback": {
            "get": {
                "tags": [
                    "Xero"
                ],
                "summary": "Handle Xero OAuth callback",
                "description": "Receives the authorization code and state from Xero after user consent and completes the OAuth connection.",
                "operationId": "07f4a41a376390811b56564befdb6896",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Authorization code returned by Xero",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "4/0AX4XfWhExampleCode"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "OAuth state used for CSRF protection",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "random_generated_state"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xero connected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xero connected successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "tenant_id": {
                                                    "type": "string",
                                                    "example": "12345678-abcd-1234-abcd-123456789abc"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Authorization code missing"
                    },
                    "422": {
                        "description": "Invalid OAuth state"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
        "/api/xero/status": {
            "get": {
                "tags": [
                    "Xero"
                ],
                "summary": "Check Xero connection status",
                "description": "Returns the current status of the Xero integration, including tenant connection details.",
                "operationId": "40ef8263ade0cb3c389c79e3a682ee63",
                "responses": {
                    "200": {
                        "description": "Xero connection status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xero connection status."
                                        },
                                        "data": {
                                            "properties": {
                                                "connected": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "tenant_id": {
                                                    "type": "string",
                                                    "example": "12345678-abcd-1234-abcd-123456789abc"
                                                },
                                                "tenant_name": {
                                                    "type": "string",
                                                    "example": "Demo Company"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
...
    },
...
}